+++ tencentcloud/extension_sqlserver.go
@@ -54,3 +54,4 @@ var SQLSERVER_DB_STATUS = map[int64]string{
 	SQLSERVER_DB_MODIFYING: "modifying",
 	SQLSERVER_DB_DELETING:  "deleting",
 }
+